support a YAML '.jekyllrc' configuration file
This patch makes the _includes location and the _layouts location
configurable, via a .jekyllrc file in the source directory.

I'm doing this because I'm maintaining a few different sites and I want
them to share _includes and _layouts.  Yes, I could use symlinks or git
submodules to duplicate the directories, but having them in one place
looks cleaner for TextMate, find, etc.